It's master instead of main according to the documentation flow.
I had an error when i was running my task with command below: for example
az acr task run --registry $ACRname --name taskhelloworld
because of main written in the url instead of master. According to the documentation flow it is master which must be there in order to have no problems for reader to execute this functionality. Otherwise user would have error ( ERROR: failed to download context. Please check  if the url is incorrect. If it has credentials. Please check if they are expired) even though user would follow instructions as written in the documentation so please update this documentation with this pull request