Recommendation to add Microsoft.AzureArcData RP
One of the most common customer support cases that we get in Arc-enabled SQL Server is when the customer has deployed the SQL Server extension to an Arc agent but the SQL Server resource doesnt show up.  Most times this is caused by the Microsoft.AzureArcData RP not being registered. If we can tell customers earlier on in the onboarding process to also register that RP as part of the onboarding process if they plan to Arc enable their SQL Servers, then hopefully we can reduce the occurrence of this type of support case.