Softening language around CIS benchmark and azcmagent
Not supported is causing customer confusion. While there are reports of incompatibilities with the CIS Benchmark, actual results may vary based on which policies the system has applied. By switching to "not tested" we more clearly indicate that we've not tried this in our lab but it may work for customers.