Improve `az acr credential-set` command documentation
The CLI documentation for `az acr credential-set` is confusing. In the
example it uses `az acr credential-set create -r MyRegistry -n MyRule`
to demonstrate how to create a credential set. Technically there is no
errors in this command, however the term `MyRule` has been used to refer
to an ACR cache rule in other examples on the page. For customer who is new
to Azure Container Registry this can be very confusing, as they started
grappling with new concepts such as _cache rule_ and _credential set_.

This commit introduces small polishes to make the documentation clearer:
- Use `MyDockerHubCredSet` when demonstrating how to operate credential
  set. This name is carefully chosen as it clearly identifies itself as
  a _credential set_, also it implies it stores credentials for DockerHub,
  and customer could create other credential sets for other registries.
- Fixed small errors on section heading, i.e. from
  `az-acr-credential-set-show` to `az acr credential-set show`.