Update arc-gateway.md
Onboarding a Windows server via Arc downloads a PowerShell script from https://gbl.his.arc.azure.com/azcmagent-windows. The downloaded script references the FQDN "aka.ms", which is not one of the 7 URLs listed in the table as requirements for the Arc Gateway, and the installation will fail due to being unable to download the Arc agent. The workaround to this is to specify the -AltDownload parameter when executing the downloaded script to use the endpoint https://gbl.his.arc.azure.com/azcmagent/latest/AzureConnectedMachineAgent.msi, which is one of the 7 allowed FQDNs and the installation will succeed. 

Note that the Linux onboarding script does not have the same requirement and does not require this modification.