Changed REST Method to PATCH
I am proposing the REST method used be changed to PATCH from of PUT because using PUT without supplying the rest of the current configuration state of the Arc-enabled Server in the JSON body will remove a server's Extended Security Update license configuration. 

Unless a user is aware of this, using the code as-is can degrade security posture. Using PUT without supplying the rest of the current configuration state can also lead to other unexpected outcomes in the future.